Test the contract per operator instead of with a sweep
Each behaviour now sits in the fixture for the operator it belongs to, where
someone changing that operator will actually see it.

QueryWhenChangedFixture, ToCollectionFixture, GroupOnFixture,
GroupImmutableFixture, SortFixture, BufferFixture, MergeChangeSetsFixture,
MergeManyChangeSetsFixture, MergeManyFixture, MaxFixture, and the four combiner
fixtures.

The recursion test uses int rather than Person because Bogus.Person and the
domain Person are both in scope in that fixture.
