+++ src/DynamicData.Tests/AggregationTests/MaxFixture.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 
 using DynamicData.Aggregation;
 using DynamicData.Tests.Domain;
@@ -6,6 +6,11 @@ using DynamicData.Tests.Domain;
 using FluentAssertions;
 
 using Xunit;
+using System.Collections.Generic;
+using System.Reactive.Concurrency;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
+using DynamicData.Binding;
 
 namespace DynamicData.Tests.AggregationTests;
 
@@ -69,5 +74,31 @@ public class MaxFixture : IDisposable
         _source.Remove("C");
         result.Should().Be(20, "Max value should be 20 after remove");
         accumulator.Dispose();
+    }
+
+    [Fact]
+    public void MaximumCompletesWhenTheSourceCompletes()
+    {
+        var completed = false;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source.Maximum(p => p.Age).Subscribe(_ => { }, () => completed = true);
+
+        source.OnCompleted();
+
+        completed.Should().BeTrue("Maximum is built on QueryWhenChanged");
+    }
+
+    [Fact]
+    public void MaximumDeliversTheErrorWithoutThrowing()
+    {
+        Exception? error = null;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source.Maximum(p => p.Age).Subscribe(_ => { }, ex => error = ex, () => { });
+
+        source.OnError(new InvalidOperationException("boom"));
+
+        error.Should().BeOfType<InvalidOperationException>();
     }
 }
