+++ src/DynamicData.Tests/List/OrFixture.cs
@@ -5,6 +5,11 @@ using System.Linq;
 using FluentAssertions;
 
 using Xunit;
+using System.Reactive.Concurrency;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
+using DynamicData.Binding;
+using DynamicData.Tests.Domain;
 
 namespace DynamicData.Tests.List;
 
@@ -136,4 +141,34 @@ public abstract class OrFixtureBase : IDisposable
     }
 
     protected abstract IObservable<IChangeSet<int>> CreateObservable();
+
+    [Fact]
+    public void CompletesOnlyWhenEverySourceCompletes()
+    {
+        var completed = false;
+
+        using var first = new Subject<IChangeSet<Person>>();
+        using var second = new Subject<IChangeSet<Person>>();
+        using var subscription = first.Or(second).Subscribe(_ => { }, () => completed = true);
+
+        first.OnCompleted();
+        completed.Should().BeFalse("the second source is still live");
+
+        second.OnCompleted();
+        completed.Should().BeTrue("every source has now finished");
+    }
+
+    [Fact]
+    public void DeliversAnErrorFromAnySource()
+    {
+        Exception? error = null;
+
+        using var first = new Subject<IChangeSet<Person>>();
+        using var second = new Subject<IChangeSet<Person>>();
+        using var subscription = first.Or(second).Subscribe(_ => { }, ex => error = ex, () => { });
+
+        second.OnError(new InvalidOperationException("boom"));
+
+        error.Should().BeOfType<InvalidOperationException>();
+    }
 }
