+++ src/DynamicData.Tests/List/QueryWhenChangedFixture.cs
@@ -1,10 +1,16 @@
-using System;
+﻿using System;
 
 using DynamicData.Tests.Domain;
 
 using FluentAssertions;
 
 using Xunit;
+using System.Collections.Generic;
+using System.Reactive;
+using System.Reactive.Concurrency;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
+using DynamicData.Binding;
 
 namespace DynamicData.Tests.List;
 
@@ -88,5 +94,31 @@ public class QueryWhenChangedFixture : IDisposable
     {
         _source.Dispose();
         _results.Dispose();
+    }
+
+    [Fact]
+    public void CompletesWhenTheSourceCompletes()
+    {
+        var completed = false;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source.QueryWhenChanged().Subscribe(_ => { }, () => completed = true);
+
+        source.OnCompleted();
+
+        completed.Should().BeTrue();
+    }
+
+    [Fact]
+    public void DeliversTheError()
+    {
+        Exception? error = null;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source.QueryWhenChanged().Subscribe(_ => { }, ex => error = ex, () => { });
+
+        source.OnError(new InvalidOperationException("boom"));
+
+        error.Should().BeOfType<InvalidOperationException>();
     }
 }
