+++ src/DynamicData.Tests/List/ToCollectionFixture.cs
@@ -3,6 +3,11 @@ using System.Collections.Generic;
 using System.Reactive.Linq;
 using FluentAssertions;
 using Xunit;
+using System.Reactive;
+using System.Reactive.Concurrency;
+using System.Reactive.Subjects;
+using DynamicData.Binding;
+using DynamicData.Tests.Domain;
 
 namespace DynamicData.Tests.List;
 
@@ -22,5 +27,33 @@ public class ToCollectionFixture
         list.Add("2");
         res1?.Count.Should().Be(2);
         res2?.Count.Should().Be(2);
+    }
+
+    [Fact]
+    public void CompletesWhenTheSourceCompletes()
+    {
+        var completed = false;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source.ToCollection().Subscribe(_ => { }, () => completed = true);
+
+        source.OnCompleted();
+
+        completed.Should().BeTrue("ToCollection is built on QueryWhenChanged");
+    }
+
+    [Fact]
+    public void ToSortedCollectionCompletesWhenTheSourceCompletes()
+    {
+        var completed = false;
+
+        using var source = new Subject<IChangeSet<Person>>();
+        using var subscription = source
+            .ToSortedCollection(SortExpressionComparer<Person>.Ascending(p => p.Name))
+            .Subscribe(_ => { }, () => completed = true);
+
+        source.OnCompleted();
+
+        completed.Should().BeTrue();
     }
 }
